Extended version of access control to events (#245)
* Extended version of access control to events

This is an extended version of access control to events based on the specs described in #235.

The user rights can be briefly described as:

Admin users:
- admin users of an organization can create event (of the default organization) with publication status PUBLIC and DRAFT
- admin users of an organization can modify and delete events of its own organizations
- admin users of an organization can modify and delete events of their sub-organizations
- admin users of an organization can modify and delete events of affiliated organization (e.g. if organization A has been assigned a responsible organization B, then the admin users of B can modify and delete events of A)

Note that the permissions between sub-organizations and affiliated organizaitons are not recursively inherited.

Regular users:
- regular users of an organization (of users' default organization ) can create event with publication status DDRAFT
- regular users of an organization can modify/delete events with publication status DRAFT

The user rights not specified above works the same way as before.

Refs: #235

* Add missing tests for some missing permission scenarios

* Replace is_admin model method for Image and Event with can_be_edited_by

This is due to regular users of an organization can also edit draft events, and this is checked on base serializer class. So we need to define the method for both Event and Image. The before for Image is the same as the old is_admin method.

* Add tests for bulk operations

* Remove duplicated old auth related tests

test_api.py contains more through tests for new permission model based on hierarchical organizations.

* Refactor: relocate event auth api to its own file

* Remove duplicated event api test