Allow auto updating datetime fields
Previously the fields `ceated_time` and `last_modified_time` are set in model's save method, and has a bug that the `created_time` is not saved when the id is explicitly assigned. This commit adds `auto_now_add` to `created_time` and `auto_now` to 'last_modified_by' to automatically save the created/modified time.

Refs: #191