Add Harrastushaku importer
Harrastushaku locations are imported as Places:

If a place in Harrastushaku has "tprek_id" set, assume it is a place that
comes from Toimipisterekisteri and don't import it from Harrastushaku.

If a place in Harrastushaku does not have "tprek_id" set, it will be imported
having data source "harrastushaku".

Harrastushaku activities are imported as Events (courses):

Activities having "active" anything else than "1" or "K" will be
ignored.

When importing and an existing course isn't present in imported data:
  - If the course's end time is in the past, the course will be left as
    it is.
  - If the course's end time is not in the past, the course will be soft
    deleted alongside its sub events.

If an activity has something in field "timetables", it will be imported
as a recurring event, otherwise as a one-time event.

A recurring course will have a super event which includes the course's
whole time period, and sub events which will represent individual course
occurrences. Other than start and end times, a super event and its sub
events will all contain the same data.

A recurring course's sub event start and end datetimes will be build using
the activity's "timetables". The time tables contain info out weekday,
times, and repetition which means number of days there is between
occurrences (basically a multiple of 7).

A recurring course's sub events will be given an ID that has the
activity's ID and start and end times of the sub event in a compressed
form. This also means that between imports only sub events that are
happening exactly at the same time are considered to be the same instance,
so if a sub event's begin or end time changes at all, a new sub event will
be created instead of updating an old one (because there is no unambiguous
way to determine which old sub event the new one corresponds to).

A course's keywords will come from both of the following:
  - The activity's main category. There are hardcoded keywords for every
    main category.
  - The activity's sub category's "searchwords". Those are manually
    entered words, which are mapped to keywords using KeywordMatcher
    (from events.keywords).

A course's audience will come from both of the following:
  - The activity's "audience_max_age" and "audience_min_age" using
    hardcoded keywords for certain age ranges.
  - The course's keywords, adding the ones that are present in
    KEYWORDS_TO_ADD_TO_AUDIENCE (from events.importer.yso).