Harrastushaku importer (#295)
* Fix a couple of issues in course extension importing

* Fix default logging configuration

* Change BaseImporter image importing

When importing data and providing it to BaseImporter.save_event(), previously
image related data were given in "image" (contained an url) and "image_license"
fields. From now on, all image related data need to be given in "images"
field, which contains a list of dicts with all image related fields, for
example:

[
    {
        "url": "www.event.com/pic1.jpg",
        "name": "pic1",
        "photographer_name": "Matti Meikäläinen",
        "license": <License Object>
    }
]

Under the hood save_event() uses new method BaseImporter.set_images(),
which accepts the same image data list as above, and can be used to just set
images instead of doing the whole save event process.

For setting an image for other than Events there is method
BaseImporter.set_image(), which accepts a single image data dict similar to
data above.

Reasons/benefits of this change:

  - all image related fields can now be imported instead of just url and license

  - nullifying images for a Place or an Event, which wasn't possible before, now
    is possible using set_image(None) and set_images([])

  - event image importing should actually work now properly, it was a bit broken
    before

* Add Harrastushaku importer

Harrastushaku locations are imported as Places:

If a place in Harrastushaku has "tprek_id" set, assume it is a place that
comes from Toimipisterekisteri and don't import it from Harrastushaku.

If a place in Harrastushaku does not have "tprek_id" set, it will be imported
having data source "harrastushaku".

Harrastushaku activities are imported as Events (courses):

Activities having "active" anything else than "1" or "K" will be
ignored.

When importing and an existing course isn't present in imported data:
  - If the course's end time is in the past, the course will be left as
    it is.
  - If the course's end time is not in the past, the course will be soft
    deleted alongside its sub events.

If an activity has something in field "timetables", it will be imported
as a recurring event, otherwise as a one-time event.

A recurring course will have a super event which includes the course's
whole time period, and sub events which will represent individual course
occurrences. Other than start and end times, a super event and its sub
events will all contain the same data.

A recurring course's sub event start and end datetimes will be build using
the activity's "timetables". The time tables contain info out weekday,
times, and repetition which means number of days there is between
occurrences (basically a multiple of 7).

A recurring course's sub events will be given an ID that has the
activity's ID and start and end times of the sub event in a compressed
form. This also means that between imports only sub events that are
happening exactly at the same time are considered to be the same instance,
so if a sub event's begin or end time changes at all, a new sub event will
be created instead of updating an old one (because there is no unambiguous
way to determine which old sub event the new one corresponds to).

A course's keywords will come from both of the following:
  - The activity's main category. There are hardcoded keywords for every
    main category.
  - The activity's sub category's "searchwords". Those are manually
    entered words, which are mapped to keywords using KeywordMatcher
    (from events.keywords).

A course's audience will come from both of the following:
  - The activity's "audience_max_age" and "audience_min_age" using
    hardcoded keywords for certain age ranges.
  - The course's keywords, adding the ones that are present in
    KEYWORDS_TO_ADD_TO_AUDIENCE (from events.importer.yso).

* Require course extension always when importing courses

* Fix Image.is_user_editable() when data_source is None