fix(api): fix failing tests
Two tests are failing due to datetime calculations that differ from the
actual implementation. The calculations in the tests don't take daylight
saving time correctly into account. This commit fixes the issue by
calculating the expected end_time in the tests in the same way as the
implementation calculates the automatic end_time.