feat(api): broaden event text search to place fields
Currently, the event API text search filters events only based on Event
fields. Thus, searching with a place name doesn't bring any results
unless the place name is included, e.g., in the location_extra_info
field of an event.

To be able to search with place names, this change supplements the
text search filters with the following place fields: name, description,
info_url, street_address, address_locality, telephone.

As a result, the text query in the API searches for the given string
both in event and place fields thus broadening the search.