remove autouse on create_initial_licenses fixture
Using *autouse* in conftest.py with a *function* scope fixture would mean
that create_initial_licenses fixture would be used for *every single test
function* that has access to conftest.py

This is problematic because then, even for a simple hello world test,
create_initial_licenses would run, which means that you need to mark
your hello world test with "django_db" marker to give it database access
so that create_initial_licenses can access the database and be able to
run.

Not all tests need to touch the database, and therefore we should not
use *autouse* on a *function* scoped fixture in conftest.py.