fix(postgres): add pg_trgm extension in the init script
The changes from linkedevents include a database migration which install
the pg_trgm extension. However, installing postgres extensions requires
superuser rights and we have restricted the user rights for the database
migration user. Thus, we can't install pg_trgm as part of the
application migrations and we need to instead install the extension as
part of the database init script which is run with superuser rights.

Note that since the init script is only run once when setting up a
database, we need to install the extension manually for any databases
that have already been initialized.