fix(make): fix importers by adding missing environment variable
Some of the importers use the SYSTEM_DATA_SOURCE_ID setting. If it isn't
set, the application uses the default value of "system". However, we
don't want to use that value so we set the environment variable
explicitly.

Note that not all of the importers or management commands use the
setting. Nonetheless, we set it to all of the importers and most of the
management commands since otherwise, the setting can easily be unset.
For instance, some of the importers have been recently updated to use
the setting but the corresponding make commands were by mistake not
updated.