fix(importer): set default name for created Language
Currently, when no languages have been created and an importer is run,
a set of languages are dynamically created based on settings.LANGUAGES
and the languages supported by the importer. However, the current
implementation sets only the "id" for the new Language and not the
"name". Thus, the created languages can't be clicked in the Django admin
list view and the only way to get to the edit view is by a direct URL.
This change fixes the issue by setting the "name" field for the Language
so that the Language can be clicked in the Django admin.