fix(make): fix the install_templates example command
The current install_templates example command doesn't actually work
since it installs the templates in the admin Docker image and not the
local development Docker image. To fix the issue, we run the command
directly on the host instead of in the admin Docker container so that
the templates are installed in the host source code directory. Since the
source code directory is mounted to the development Docker container,
running the install_templates command will correctly install the
templates for the local development environment.