refactor(helmet): sort the library listing based on city
Sort the library listing / mapping in the helmet importer based on city
so that it's easier to see what libraries are missing (Helmet.fi lists
the libraries also based on city).

Also, add a note about the library location mapping not being
up-to-date. An up-to-date list of libraries and their IDs should be
requested from the Helmet maintainer so that it can be updated to
espooevents-service / Linked Events.