feat(make): change the install_templates command to be run in container
Change the install_templates command to be run in a Docker container.
This was previously changed to be run on the host since the previous
command that executed the command in a Docker container didn't work.
However, the command can easily be fixed to work in a Docker container
by just mounting the source code directory to the container where the
install_templates command is being run. This way, the installed
templates are correctly installed to the correct host source code
directory which is also visible for the container that runs
espooevents-service.