fix(helmet): fix automatic creation of Internet location
Currently, the helmet importer checks if an "Internet" location exists
and creates one if it doesn't exist. This, however, currently fails since
the "self.system_data_source" variable passed as an argument to the Place
creation query contains a tuple and not a DataSource object.

This change fixes the bug by correctly assigning the DataSource object
to the "self.system_data_source" variable so that the "Internet" location
can be successfully created if it doesn't already exist.