LINK-485. Fix failing PUT requests
Querying *.objects.all() for related fields results in huge requests whenever an Event. Internal DRF mechanics for related fields as specified in relations.py uses get() for every updated related field and presumes an existing queryset. The goal of this bug fix is to provide querysets for the related fields that would contain only the related objects mentioned in the request