Set permission for development env MEDIA_ROOT mount
Changes the permissions for the local named volume mount point
django-media-volume. Current Dockerfile uses a non-root user. Locally this
doesn't work with the named volume (on Linux at least) since the container
user doesn't have permission to	write into the volume. Volume will
however inherit the permission of a pre-existing folder	(which might
be a bit hacky) on docker so this is a slight workaround.

Refs LINK-1233