Personal information fields for an external user are required
Admin and regular users (members of an organization) are not required
to fill the personal information fields.

For an external user creator_name and maximum_attendee_capacity fields are
required. Either user_email or user_phone_number has to be filled.

If any personal information fields are filled user_consent is required.

Refs LINK-1139