User doing AD authentication is never considered to be an external user
User.is_external checks the AMR claim value and and returns False if
AD authentication is used.

AMR (Authentication Method Reference) should be a list of string
that are identifiers for authentication methods used in the authentication.
Tunnistamo however sets a simple string here, which is assumed in
this implementation also.

Refs LINK-1394