Remove use of distinct in event listing
The use of distinct is extremely slow. It's use can be avoided in
one-to-many and many-to-many relations by using
exclude(~Q(relation=something). When excluding relations, Django
automatically uses SQL EXISTS. Negating the Q statement will
essentially create a NOT NOT EXISTS which the database then optimizes
into plain EXISTS.

This appears to be still notably faster than using
filter(Exists(subquery)) which creates an INNER JOIN inside the exists
clause.

Ref LINK-1408