fix: disable user created email notification
Currently an email notification is sent to all superusers about every
new user. Now that we have outside users and registration users, we most
probably don't want to be sending those notifications anymore. The
notification is just disabled for now, but should be completely removed
later if there is no use for it.