fix(events): fix error with x_ongoing_OR_set
If no events were found, the filter caused a crash
(TypeError: intersection needs values). This fixes
the issue. This also fixes another issue; this
error was encountered with the query param
"local_ongoing_OR_set=xxxxxxx", which shouldn't
even trigger the filter since it requires an
ascending number at the end, starting from 1.
E.g. "OR_set1=x,OR_set2=y" is valid.

There's also something terribly wrong with
the regex that the ongoing OR set filters
use. Apparently, this was encountered in f581b16a,
since there are tests marked as xfail for the
function.

Added some test cases for the filter (there were
none before), but marked them as xfail because
the filter doesn't seem to work as described
in documentation.

Refs LINK-1986