fix(registration): order signups by id
Order signups by the id in the signup endpoint to achieve correct
ordering in the UI for the attendee list and waitlist listing.

Reason for use the ID for ordering is that the waitlist logic uses the
id for ordering waitlist queue. Other solution would be to change it
to be using created_time.

Refs LINK-2187