fix(events): use django qs functions in max_* and min_duration filters
There were 'ProgrammingError' thrown when using max or min_duration
filter with hide_recurring_children filter. This was because the
*_duration filters were defined to use custom sql WHERE clauses without
referring to which table the fields in the WHERE clause was referring to
and there were other tables joined to the query by hide_recurring_
children filter.

Solution was to convert these max and min_duration filters to be using
django qs functions.

Also refactors the max_ and min_duration filters away from api.py to
filters.

Refs LINK-2203