chore(registrations): send event cancellation notifications in order
This was causing failures with two tests which assume that the
notifications were sent in a certain order (desc pk). However,
if an order is not specified, Postgres cannot guarantee an order,
and that's exactly what happened here.

There were two ways of resolving this; either guarantee an order or
change the test to be order-agnostic. In this case, I chose the
path of least resistance.

Refs: LINK-2210