fix: add grace period to seat reservations expiration
Seat reservation expiry time was determined purely from reservation
start time and the number of seats. This fix changes the logic and adds
an explicit expiration field that is adjusted when the number of seats
is changed. The logic is now as follows:

1) When increasing the number of seats, expiration time will always
increase.
2) When decreasing the number of seats, expiration time will either
remain the same if it was already under the grace period or it will be
decreased, but never below the defined grace period.

Grace period defaults to 5 minutes.

refs: LINK-2234