feat: use word bases instead of syllables
Analyze words with Voikko and use
word bases directly instead of syllables.

This will result into direct parsing of
the finnish words in nominative form
even for compound words,
which is a desired goal here.

Do the same for the search input
to improve the hit rate.

Also updated tests.

Refs: LINK-2270