feat: proxy image url
This adds a new query parameter `use_image_proxy` to event and
image endpoints. When set to true, image urls will be rewritten via
a new image-file endpoint. This endpoint returns a HTTP redirect
to the actual image file.

This is to be used in conjunction with a web server that will follow
the redirect and then return the image data directly from the same
domain.

This in turn allows for stricter CSP configuration.

refs: LINK-2355